+++ src/com/google/lecture_manager/server/jdbc/dao/UserDAO.java
@@ -105,7 +105,7 @@ public class UserDAO {
       statement.setLong(1, UserTypes.TEACHER.getId());
       rs = statement.executeQuery();
       while (rs.next()){
-        result.add(new TeacherDTO(rs.getLong(1), rs.getString(2), rs.getString(3), rs.getString(5), rs.getString(4)));
+        result.add(new TeacherDTO(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getString(5), rs.getString(4)));
       }
       return result;
     } finally {
